fix(cmd): Split gontc command to several files
